+++ packages/client/hmi-client/src/components/workflow/ops/stratify-mira/mod.ts
@@ -1,6 +1,6 @@
 import { StratifyMiraOperation as operation } from './stratify-mira-operation';
-import node from './tera-stratify-node-mira.vue';
-import drilldown from './tera-stratify-mira.vue';
+import node from './tera-stratify-node.vue';
+import drilldown from './tera-stratify-drilldown.vue';
 
 const name = operation.name;
